:technologist: Automatically inject nano newlib for ARM baremetal (#41)
The only place in libhal where we require injecting nano newlib is with
test packages as they are generic. This will ensure that test packages
link automatically. These binaries may not work correctly if a library
uses picolibc and they intersect in ways that break the code. But
thats not a concern currently. Demos should be the location for testing
this.